+++ libr/anal/Makefile
@@ -31,7 +31,7 @@ STATIC_OBJS=$(addprefix $(LTOP)/anal/p/,$(STATIC_OBJ))
 OBJLIBS=meta.o reflines.o op.o fcn.o bb.o var.o
 OBJLIBS+=cond.o value.o cc.o class.o diff.o type.o
 OBJLIBS+=hint.o anal.o data.o xrefs.o esil.o sign.o
-OBJLIBS+=anal_ex.o switch.o state.o cycles.o
+OBJLIBS+=anal_ex.o switch.o state.o cycles.o esil_dfg.o
 OBJLIBS+=esil_sources.o esil_interrupt.o esil_cfg.o
 OBJLIBS+=esil_stats.o esil_trace.o flirt.o labels.o
 OBJLIBS+=esil2reil.o pin.o session.o vtable.o rtti.o
